requirements.txt
asgiref==3.11.0
Django==5.2.10
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.2
django-cors-headers==4.3.1
